+++ lib/custom_picker/history_colors.dart
@@ -108,7 +108,10 @@ class _HistoryColorsState extends State<HistoryColors> {
                   padding: const EdgeInsets.only(left: 27, right: 17),
                   crossAxisCount: 5,
                   mainAxisSpacing: 12,
-                  crossAxisSpacing: ((maxWidth(context) - 304) / 5),
+                  crossAxisSpacing: ((maxWidth(context) - 304) / 5).clamp(
+                    0.0,
+                    double.infinity,
+                  ),
                   dragStartBehavior: DragStartBehavior.down,
                   children: List.generate(
                     historyColors.length >= 10